% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseIsEncryptionOn.R
\name{databaseIsEncryptionOn}
\alias{databaseIsEncryptionOn}
\title{Database Encryption Status}
\arguments{
\item{conn}{Connection to an existing database to be checked.}
}
\value{
TRUE if the database as encryption enabled, or FALSE otherwise.
}
\description{
Checks encryption status of an existing database.
}
\examples{
  key <- "0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF"
  tmp_file <- tempfile()
  con <- dbConnect(RSQLCipher::SQLCipher(), tmp_file, key = key)

  databaseIsEncryptionOn(con)  # Expect TRUE

  dbDisconnect(con)

}
