% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseKeyRemove.R
\name{databaseKeyRemove}
\alias{databaseKeyRemove}
\title{databaseKeyRemove}
\arguments{
\item{conn}{Connection to an existing encrypted database to be decrypted.}

\item{file}{Optional, path to the new decrypted database. A temporary file will be generated if
not provided.}
}
\value{
A named list with True/False and the file name of the plain database
}
\description{
Creates an unencrypted/plain copy of an existing encrypted database.
\emph{THIS FUNCTION MAY TAKE LONG FOR BIG DATABASE FILES}
}
\examples{
  key <- "0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF"
  tmp_file <- tempfile()
  con <- dbConnect(RSQLCipher::SQLCipher(), tmp_file, key = key)
  dbWriteTable(con, "mtcars", mtcars)

  newDB <- databaseKeyRemove(con)
  print(newDB)
  dbDisconnect(con)

  con2 <- dbConnect(RSQLCipher::SQLCipher(), newDB$file)
  dbListTables(con2)
  dbDisconnect(con2)

}
