% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseKeyChange.R
\name{dbChangeKey}
\alias{dbChangeKey}
\title{dbChangeKey
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
dbChangeKey(conn, old_key = NULL, new_key = NULL)
}
\arguments{
\item{conn}{database Connection}

\item{old_key}{Old database key, a character string of size 64, containing
32 hex encoded characters currently used as key for the database.
encryption.}

\item{new_key}{New database key, a character string of size 64, containing
32 hex encoded characters to be used as new key for the database.}
}
\description{
Changes the key for an encrypted database.
}
\seealso{
\code{\link[=databaseKeyChange]{databaseKeyChange()}}
}
