% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{execute}
\alias{execute}
\title{Execute a SQL query on a table and return a data frame}
\usage{
execute(df, enforce.types = F)
}
\arguments{
\item{df}{A table or lazy query, returned from \link{load_table} and
\link[dplyr]{dplyr} functions.}

\item{enforce.types}{Should column types be inferred from the input types,
and be enforced when the data frame is constructed from the SQL output.}
}
\value{
A \link[tibble]{tibble} containing the queried data.
}
\description{
Execute a SQL query on a table and return a data frame
}
