% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_sql}
\alias{run_sql}
\title{Execute a SQL query on a database and return the raw text result.}
\usage{
run_sql(path, query)
}
\arguments{
\item{path}{A character vector containing the path to the database.}

\item{query}{A character vector containing the SQL code to run.}
}
\value{
A character vector, with each element a line of output.
}
\description{
Will prompt for the database key if not stored in SQL_KEY environment variable.
}
