% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseKeyChange.R
\name{databaseKeyChange}
\alias{databaseKeyChange}
\title{databaseKeyChange}
\arguments{
\item{conn}{database Connection}

\item{old_key}{Old database key, a character string of size 64, containing
32 hex encoded characters currently used as key for the database.
encryption.}

\item{new_key}{New database key, a character string of size 64, containing
32 hex encoded characters to be used as new key for the database.}
}
\value{
True if successful or False otherwise#'
}
\description{
Changes the key for an encrypted database.
}
\examples{
  key_1 <- "0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF"
  key_2 <- "ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789"
  tmp_file <- tempfile()
  con <- dbConnect(RSQLCipher::SQLCipher(), tmp_file, key = key_1)

  databaseKeyChange(con, key_1, key_2)
  dbDisconnect(con)

  con2 <- dbConnect(RSQLCipher::SQLCipher(), tmp_file, key = key_2)

}
