% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbChangeKey.R
\name{dbChangeKey}
\alias{dbChangeKey}
\title{dbChangeKey}
\arguments{
\item{conn}{database Connection}

\item{old_key}{Old database key, a character string of size 64, containing
32 hex encoded characters currently used as key for the database.
encryption.}

\item{new_key}{New database key, a character string of size 64, containing
32 hex encoded characters to be used as new key for the database.}
}
\description{
Changes the key for an encrypted database.
}
